requirements.txt
flask
pillow
pymongo
scikit-learn
Tensorflow
Keras
face_recognition